requirements.txt
glob2==0.6
imageio==2.3.0
imageio-ffmpeg==0.4.5
imagesize==1.0.0
imbalanced-learn==0.8.1
imblearn==0.0
Keras==2.2.4
matplotlib==2.2.2
numpy==1.14.3
numpydoc==0.8.0
opencv-python==4.5.4.60
pandas==0.23.0
Pillow==5.1.0
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.24.2
scipy==1.1.0
seaborn==0.8.1
statsmodels==0.9.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
toolz==0.9.0
torch==1.4.0
torchvision==0.5.0
tornado==5.0.2
tqdm==4.62.3
unicodecsv==0.14.1
urllib3==1.22